requirements.txt
streamlit
pandas
openpyxl
pillow
matplotlib